[feat] #34 git rm 기능 구현

+++ src/main/java/com/github/filemanager/FileManager.java
@@ -1062,11 +1062,11 @@ public class FileManager {
 
     --> 1.버튼이 눌렸을 떄 선택된 파일이 있는지 확인 (o)
     --> 2.git repo안의 파일인지 확인 (o)
-    --> 3.git rm 명령어 실행 (o)
-    --> 4.결과에 따른 반환값에 따른 팝업창 생성 (o)
+    --> 3.해당 파일이 Committed 상태이거나 Unmodified상태인지 확인 (o).
+    --> 4.git rm 명령어 실행 (o)
+    --> 5.결과에 따른 반환값에 따른 팝업창 생성 (o)
     */
 
-        //
         if (currentFile == null) {//1.파일 선택하지 않았을 경우, 별도행위없이 함수종료
             showErrorMessage("파일을 선택해주세요.", "Select File");
             return;
@@ -1074,32 +1074,34 @@ public class FileManager {
 
         if(isFileInGitRepository()) {//2.git repo안에 있는 경우에만 실행.
             try{
-                int result = JOptionPane.showConfirmDialog(gui, "해당 파일을 삭제하고 이 변화를 staged하시겠습니까?", "git rm", JOptionPane.ERROR_MESSAGE);
+                if(isCommittedOrUnmodifiedFile(currentFile)) {//파일이 Committed, Unmodified 상태인 경우에만 실행
+                    int result = JOptionPane.showConfirmDialog(gui, "해당 파일을 삭제하고 이 변화를 staged하시겠습니까?", "git rm", JOptionPane.ERROR_MESSAGE);
 
-                if (result == JOptionPane.OK_OPTION) { // "예" 클릭 시 git rm 명령어 실행
-                    String[] gitRmCommand = {"git", "rm", currentFile.getName()};
-                    ProcessBuilder processBuilder = new ProcessBuilder(gitRmCommand);
-                    processBuilder.directory(currentFile.getParentFile());
-                    Process process = processBuilder.start();
-                    int rmStatus = process.waitFor(); //git rm 명령어 정상 실행 여부
+                    if (result == JOptionPane.OK_OPTION) { // "예" 클릭 시 git rm 명령어 실행
+                        String[] gitRmCommand = {"git", "rm", currentFile.getName()};
+                        ProcessBuilder processBuilder = new ProcessBuilder(gitRmCommand);
+                        processBuilder.directory(currentFile.getParentFile());
+                        Process process = processBuilder.start();
+                        int rmStatus = process.waitFor(); //git rm 명령어 정상 실행 여부
 
-                    if (rmStatus == 0){ // git rm 명령어가 정상적으로 실행되어 status가 0일 경우
-                        JOptionPane.showMessageDialog(gui, "성공적으로 파일을 remove 했습니다.");
-                        System.out.println(currentFile);
-                        System.out.println("removed && change staged");
-                        try{
-                            renderGitFileStatus(); //스테이지했을 경우, 파일에 변화가 일어났으므로 렌더링
-                            TreePath parentPath = findTreePath(currentFile.getParentFile());
-                            DefaultMutableTreeNode parentNode = (DefaultMutableTreeNode) parentPath.getLastPathComponent();
-                            showChildren(parentNode);
-                        }catch (IOException | GitAPIException e){
-                            e.printStackTrace();
+                        if (rmStatus == 0) { // git rm 명령어가 정상적으로 실행되어 status가 0일 경우
+                            JOptionPane.showMessageDialog(gui, "성공적으로 파일을 remove 했습니다.");
+                            System.out.println(currentFile);
+                            System.out.println("removed && change staged");
+                            try {
+                                renderGitFileStatus(); //스테이지했을 경우, 파일에 변화가 일어났으므로 렌더링
+                                TreePath parentPath = findTreePath(currentFile.getParentFile());
+                                DefaultMutableTreeNode parentNode = (DefaultMutableTreeNode) parentPath.getLastPathComponent();
+                                showChildren(parentNode);
+                            } catch (IOException | GitAPIException e) {
+                                e.printStackTrace();
+                            }
+                        } else { //git rm 명령어가 정상적으로 실행되지 않았을 경우
+                            showErrorMessage("파일을 remove하는 과정에서 오류가 발생했습니다.", "git rm error");
                         }
-                    }else{ //git rm 명령어가 정상적으로 실행되지 않았을 경우
-                        showErrorMessage("파일을 remove하는 과정에서 오류가 발생했습니다.","git rm error");
                     }
+                    gui.repaint();
                 }
-                gui.repaint();
             } catch (InterruptedException | IOException e){
                 e.printStackTrace();
             }
@@ -1158,6 +1160,27 @@ public class FileManager {
             e.printStackTrace();
         }
 
+    }
+    /**
+     * 단일 파일을 선택 했을 때 해당 파일이 Commited or UnModified 상태인지 확인해 주는 boolean 함수
+     */
+    private boolean isCommittedOrUnmodifiedFile(File file){
+        try{
+            Git git;
+            git = Git.open(currentFile.getParentFile());
+            Status status = git.status().call();
+
+            Set<String> untracked = status.getUntracked();  //Untracked 파일 이름을 받아와 비교
+
+            //untracked, modified, staged가 아니라면 Committed or Unmodified상태.
+            if (!untracked.contains(file.getName()) && !isModifiedFile(file) && !isStagedFile(file)){return true;}
+            else{
+                showErrorMessage("선택한 파일은 Committed나 UnModified 상태가 아닙니다. ", "Committed or Unmodified file chosen error");
+            }
+        }catch (IOException | GitAPIException e ){
+            e.printStackTrace();
+        }
+        return false;
     }
 
     /**
